Created custom Profile Screen Button, Separated style into its own file
