+++ App.js
@@ -1,11 +1,13 @@
 import { StyleSheet, SafeAreaView, StatusBar } from 'react-native';
 import Navigation from './src/navigation';
+import ProfilePage from './src/screens/ProfileScreen';
 export default function App() {
   return (
     <SafeAreaView style={styles.container}>
       <StatusBar />
       {/* Navigation will go here */}
-      <Navigation />
+      {/*<Navigation />*/}
+      <ProfilePage/>
     </SafeAreaView>
   );
 }
