+++ src/components/ProfileScreenButton/styles.js
@@ -1,0 +1,26 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+      flex: .5,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    button: {
+        alignItems: 'center',
+        padding:10,
+        backgroundColor: '#99182e',
+        borderRadius: 4,
+        paddingVertical:12,
+    },
+    text:{
+        color: 'white',
+    },
+    space:{
+        width:20,
+        height:20
+    }
+  });
+
+  export default styles;
