+++ src/screens/ProfileScreen/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    space:{
+        width:20,
+        height:20
+    }
+  });
+
+  export default styles;
